refactor and move snapshotmanager
