+++ tests/integration/awslambda/test_lambda.py
@@ -237,7 +237,7 @@ class TestLambdaAPI:
             runtime=LAMBDA_RUNTIME_PYTHON36,
         )
         lambda_arn = lambda_create_response["CreateFunctionResponse"]["FunctionArn"]
-        snapshot.assert_match("create_lambda", lambda_create_response)
+        snapshot.match("create_lambda", lambda_create_response)
         # create lambda permission
         action = "lambda:InvokeFunction"
         sid = "s3"
@@ -249,11 +249,11 @@ class TestLambdaAPI:
             Principal=principal,
             SourceArn=aws_stack.s3_bucket_arn("test-bucket"),
         )
-        snapshot.assert_match("add_permission", resp)
+        snapshot.match("add_permission", resp)
 
         # fetch lambda policy
         get_policy_result = lambda_client.get_policy(FunctionName=function_name)
-        snapshot.assert_match("get_policy", get_policy_result)
+        snapshot.match("get_policy", get_policy_result)
         assert lambda_arn == json.loads(get_policy_result["Policy"])["Statement"][0]["Resource"]
 
     # TODO permissions cannot be added to $LATEST
@@ -328,7 +328,7 @@ class TestLambdaAPI:
             StatementId=sid,
             Principal=principal,
         )
-        snapshot.assert_match("add_permission_1", permission_1_add)
+        snapshot.match("add_permission_1", permission_1_add)
 
         sid_2 = "sqs"
         principal_2 = "sqs.amazonaws.com"
@@ -339,11 +339,11 @@ class TestLambdaAPI:
             Principal=principal_2,
             SourceArn=aws_stack.s3_bucket_arn("test-bucket"),
         )
-        snapshot.assert_match("add_permission_2", permission_2_add)
+        snapshot.match("add_permission_2", permission_2_add)
         policy_response = lambda_client.get_policy(
             FunctionName=function_name,
         )
-        snapshot.assert_match("policy_after_2_add", policy_response)
+        snapshot.match("policy_after_2_add", policy_response)
 
         with pytest.raises(ClientError) as e:
             lambda_client.remove_permission(
@@ -361,7 +361,7 @@ class TestLambdaAPI:
                 FunctionName=function_name,
             )["Policy"]
         )
-        snapshot.assert_match("policy_after_removal", policy)
+        snapshot.match("policy_after_removal", policy)
         assert policy["Statement"][0]["Sid"] == sid
 
         lambda_client.remove_permission(
@@ -464,7 +464,7 @@ class TestLambdaAPI:
             MaximumEventAgeInSeconds=123,
             DestinationConfig=destination_config,
         )
-        snapshot.assert_match("put_function_event_invoke_config", response)
+        snapshot.match("put_function_event_invoke_config", response)
 
         # over writing event invoke config
         response = lambda_client.put_function_event_invoke_config(
@@ -472,14 +472,14 @@ class TestLambdaAPI:
             MaximumRetryAttempts=2,
             DestinationConfig=destination_config,
         )
-        snapshot.assert_match("put_function_event_invoke_config_overwritemaxeventage", response)
+        snapshot.match("put_function_event_invoke_config_overwritemaxeventage", response)
 
         # updating event invoke config
         response = lambda_client.update_function_event_invoke_config(
             FunctionName=function_name,
             MaximumRetryAttempts=1,
         )
-        snapshot.assert_match("put_function_event_invoke_config_maxattempt1", response)
+        snapshot.match("put_function_event_invoke_config_maxattempt1", response)
 
         # clean up
         lambda_client.delete_function_event_invoke_config(FunctionName=function_name)
@@ -496,10 +496,10 @@ class TestLambdaAPI:
         response = lambda_client.put_function_concurrency(
             FunctionName=function_name, ReservedConcurrentExecutions=123
         )
-        snapshot.assert_match("put_function_concurrency", response)
+        snapshot.match("put_function_concurrency", response)
         assert "ReservedConcurrentExecutions" in response
         response = lambda_client.get_function_concurrency(FunctionName=function_name)
-        snapshot.assert_match("get_function_concurrency", response)
+        snapshot.match("get_function_concurrency", response)
         assert "ReservedConcurrentExecutions" in response
         lambda_client.delete_function_concurrency(FunctionName=function_name)
 
@@ -522,7 +522,7 @@ class TestLambdaAPI:
             CodeSigningPolicies={"UntrustedArtifactOnDeployment": "Enforce"},
         )
         snapshot.replace_value(re.compile(r"^csc-[0-9a-f]{17}$"), "<csc-id>")
-        snapshot.assert_match("create_code_signing_config", response)
+        snapshot.match("create_code_signing_config", response)
 
         assert "Description" in response["CodeSigningConfig"]
         assert "SigningProfileVersionArns" in response["CodeSigningConfig"]["AllowedPublishers"]
@@ -535,7 +535,7 @@ class TestLambdaAPI:
             CodeSigningConfigArn=code_signing_arn,
             CodeSigningPolicies={"UntrustedArtifactOnDeployment": "Warn"},
         )
-        snapshot.assert_match("update_code_signing_config", response)
+        snapshot.match("update_code_signing_config", response)
 
         assert (
             "Warn"
@@ -543,17 +543,17 @@ class TestLambdaAPI:
         )
         response = lambda_client.get_code_signing_config(CodeSigningConfigArn=code_signing_arn)
         assert 200 == response["ResponseMetadata"]["HTTPStatusCode"]
-        snapshot.assert_match("get_code_signing_config", response)
+        snapshot.match("get_code_signing_config", response)
 
         response = lambda_client.put_function_code_signing_config(
             CodeSigningConfigArn=code_signing_arn, FunctionName=function_name
         )
         assert 200 == response["ResponseMetadata"]["HTTPStatusCode"]
-        snapshot.assert_match("put_function_code_signing_config", response)
+        snapshot.match("put_function_code_signing_config", response)
 
         response = lambda_client.get_function_code_signing_config(FunctionName=function_name)
         assert 200 == response["ResponseMetadata"]["HTTPStatusCode"]
-        snapshot.assert_match("get_function_code_signing_config", response)
+        snapshot.match("get_function_code_signing_config", response)
         assert code_signing_arn == response["CodeSigningConfigArn"]
         assert function_name == response["FunctionName"]
 
@@ -580,7 +580,7 @@ class TestLambdaAPI:
             StatementId=sid,
             Principal="logs.amazonaws.com",
         )
-        snapshot.assert_match("add_permission_response_1", resp)
+        snapshot.match("add_permission_response_1", resp)
         assert "Statement" in resp
 
         sid = "kinesis"
@@ -590,14 +590,14 @@ class TestLambdaAPI:
             StatementId=sid,
             Principal="kinesis.amazonaws.com",
         )
-        snapshot.assert_match("add_permission_response_2", resp)
+        snapshot.match("add_permission_response_2", resp)
 
         assert "Statement" in resp
 
         policy_response = lambda_client.get_policy(
             FunctionName=function_name,
         )
-        snapshot.assert_match("policy_after_2_add", policy_response)
+        snapshot.match("policy_after_2_add", policy_response)
 
 
 class TestLambdaBaseFeatures:
@@ -625,7 +625,7 @@ class TestLambdaBaseFeatures:
             DeadLetterConfig={"TargetArn": queue_arn},
             role=lambda_su_role,
         )
-        snapshot.assert_match("create_lambda_with_dlq", create_lambda_response)
+        snapshot.match("create_lambda_with_dlq", create_lambda_response)
 
         # invoke Lambda, triggering an error
         payload = {lambda_integration.MSG_BODY_RAISE_ERROR_FLAG: 1}
@@ -643,7 +643,7 @@ class TestLambdaBaseFeatures:
             assert "RequestID" in msg_attrs
             assert "ErrorCode" in msg_attrs
             assert "ErrorMessage" in msg_attrs
-            snapshot.assert_match("sqs_dlq_message", result)
+            snapshot.match("sqs_dlq_message", result)
 
         # on AWS, event retries can be quite delayed, so we have to wait up to 6 minutes here, potential flakes
         retry(receive_dlq, retries=120, sleep=3)
@@ -652,14 +652,14 @@ class TestLambdaBaseFeatures:
         update_function_config_response = lambda_client.update_function_configuration(
             FunctionName=lambda_name, DeadLetterConfig={}
         )
-        snapshot.assert_match("delete_dlq", update_function_config_response)
+        snapshot.match("delete_dlq", update_function_config_response)
         # invoke Lambda again, assert that status code is 200 and error details contained in the payload
         result = lambda_client.invoke(
             FunctionName=lambda_name, Payload=json.dumps(payload), LogType="Tail"
         )
         result = read_streams(result)
         payload = json.loads(to_str(result["Payload"]))
-        snapshot.assert_match("result_payload", payload)
+        snapshot.match("result_payload", payload)
         assert 200 == result["StatusCode"]
         assert "Unhandled" == result["FunctionError"]
         assert "$LATEST" == result["ExecutedVersion"]
@@ -712,7 +712,7 @@ class TestLambdaBaseFeatures:
                 "OnFailure": {"Destination": queue_arn},
             },
         )
-        snapshot.assert_match("put_function_event_invoke_config", put_event_invoke_config_response)
+        snapshot.match("put_function_event_invoke_config", put_event_invoke_config_response)
 
         lambda_client.invoke(
             FunctionName=lambda_name,
@@ -726,7 +726,7 @@ class TestLambdaBaseFeatures:
             msg = rs["Messages"][0]["Body"]
             msg = json.loads(msg)
             assert condition == msg["requestContext"]["condition"]
-            snapshot.assert_match("destination_message", rs)
+            snapshot.match("destination_message", rs)
 
         retry(receive_message, retries=120, sleep=3)
 
@@ -744,7 +744,7 @@ class TestLambdaBaseFeatures:
         payload_bytes = to_bytes(json.dumps(payload))
         result = lambda_client.invoke(FunctionName=function_name, Payload=payload_bytes)
         result = read_streams(result)
-        snapshot.assert_match("invocation_response", result)
+        snapshot.match("invocation_response", result)
         assert 200 == result["ResponseMetadata"]["HTTPStatusCode"]
         result_data = result["Payload"]
         result_data = json.loads(to_str(result_data))
@@ -782,7 +782,7 @@ class TestPythonRuntimes:
         snapshot.register_replacement(
             re.compile(r"\d{4}/\d{2}/\d{2}/\[((\$LATEST)|\d+)\][0-9a-f]{32}"), "<log_stream_id>"
         )
-        snapshot.assert_match("invoke", result)
+        snapshot.match("invoke", result)
         result_data = json.loads(result["Payload"])
 
         # assert response details
@@ -796,7 +796,7 @@ class TestPythonRuntimes:
             re.compile(r"Duration: \d+(\.\d{2})? ms"), "Duration: <duration> ms"
         )
         snapshot.register_replacement(re.compile(r"Used: \d+ MB"), "Used: <memory> MB")
-        snapshot.assert_match("logs", {"logs": logs})
+        snapshot.match("logs", {"logs": logs})
         assert "START" in logs
         assert "Lambda log message" in logs
         assert "END" in logs
